+++ tests/ops/integration_tests/saas/test_snap_task.py
@@ -1,11 +1,10 @@
 import pytest
 
 from fides.api.models.policy import Policy
-from fides.config import CONFIG
 from tests.ops.integration_tests.saas.connector_runner import ConnectorRunner
 
 
-# @pytest.mark.skip(reason="Currently unable to test OAuth2 connectors")
+@pytest.mark.skip(reason="Currently unable to test OAuth2 connectors")
 @pytest.mark.integration_saas
 class TestSnapConnector:
     def test_connection(self, snap_runner: ConnectorRunner):